itoro/ai_crypto_agents/requirements.txt
# Core Python Dependencies for ITORO AI Crypto Agents
# =================================================================

# Web Framework & API
Flask==3.0.0
fastapi==0.104.1
uvicorn[standard]==0.24.0
requests==2.31.0
gunicorn==21.2.0

# Database & Data Processing
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pandas==2.1.4
numpy==1.26.2
sqlite3

# AI/ML Models & APIs
openai==1.3.8
anthropic==0.7.8
google-generativeai==0.3.2
groq==0.4.1

# Data Analysis & Technical Indicators
ta-lib==0.4.28
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0

# Cryptocurrency & Blockchain
solana==0.32.0
solders==0.20.4
web3==6.11.4
eth-account==0.9.0

# APIs & Integrations
ccxt==4.1.63
python-binance==1.0.19
coinbase-advanced-py==1.2.0
helius==0.1.0

# Real-time Data & Messaging
redis==5.0.1
pika==1.3.2
websockets==12.0

# GUI & Desktop Application
PySide6==6.6.1
PyQt6==6.6.1

# Configuration & Environment
python-dotenv==1.0.0
pyyaml==6.0.1

# Utilities & Helpers
termcolor==2.4.0
colorama==0.4.6
tqdm==4.66.1
schedule==1.2.1
 APScheduler==3.10.4

# Date & Time Handling
python-dateutil==2.8.2

# Logging & Monitoring
loguru==0.7.2

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Optional Dependencies (may be installed separately)
# =================================================================

# Supabase (for cloud database integration)
# supabase==2.3.0

# Additional AI Models (optional)
# transformers==4.35.2
# torch==2.1.1
# accelerate==0.24.1

# Advanced Analytics (optional)
# statsmodels==0.14.0
# scipy==1.11.4

# Jupyter (for analysis notebooks)
# jupyter==1.0.0
# notebook==7.0.6

# Ngrok (for webhook tunneling in development)
# pyngrok==7.0.3

# Telegram Bot (for notifications)
# python-telegram-bot==20.7

# Discord Bot (for notifications)
# discord.py==2.3.2

# Twitter/X API (for social sentiment)
# tweepy==4.14.0
